requirements.txt
Django>=1.5,<1.6
django-tastypie==0.9.12
Fabric==1.4.0
gevent==0.13.6
greenlet==0.3.4
gunicorn==0.13.4
psycopg2==2.5
pycrypto==2.5
ssh==1.7.13
wsgiref==0.1.2
django-pylibmc-sasl==0.2.4
pylibmc==1.2.2
geopy
django-debug-toolbar
South

